+++ components/QuestionnaireExtractTest.vue
@@ -1,10 +1,17 @@
 <template>
   <div class="ace_editor">
+    <v-text-field label="Extract Service URL" v-model="extractServiceUrl" :loading="extractingInProgress">
+      <template v-slot:append>
+        <v-btn @click="performExtractOperation">
+          Extract!
+        </v-btn>
+      </template>
+    </v-text-field>
     <div height="85px" width="100%" ref="aceEditorExpression"></div>
-    <v-textarea auto-grow :rows="5" label="Field"  />
     <div class="ace_editor_footer"></div>
-    <v-textarea auto-grow :rows="5" label="Field" :value="JSON.stringify(questionnaire, null, 2)" />
-    <v-textarea auto-grow :rows="5" label="Field" :value="JSON.stringify(questionnaireResponse, null, 2)" />
+    <v-textarea :rows="5" small label="Extract Parameters" :value="JSON.stringify(extractParameters, null, 2)" />
+    <v-textarea :rows="5" small label="Questionnaire" :value="JSON.stringify(questionnaire, null, 2)" />
+    <v-textarea :rows="5" small label="QuestionnaireResponse" :value="JSON.stringify(questionnaireResponse, null, 2)" />
   </div>
 </template>
 
@@ -23,8 +30,8 @@
 </style>
 
 <script lang="ts">
-import { Questionnaire, QuestionnaireResponse } from "fhir/r4b";
-import { Component, Prop, Vue } from "vue-property-decorator";
+import { Parameters, Questionnaire, QuestionnaireResponse } from "fhir/r4b";
+import { Component, Prop, Vue, Watch } from "vue-property-decorator";
 
 import "ace-builds";
 import ace from "ace-builds";
@@ -38,59 +45,123 @@ import "~/assets/fhirpath_highlighter.scss"
 @Component
 export default class QuestionnaireExtractTest extends Vue {
 
+  // QuestionnaireExtractTest(){
+  //   this.extractParameters = "Extracting the questionnaire";
+  // }
+
   // Properties provided by the parent component
   @Prop() readonly questionnaire: Questionnaire | undefined;
   @Prop() readonly questionnaireResponse: QuestionnaireResponse | undefined;
 
   // Properties visible to the local template
   public selectedTabValue: number = 2;
+  public extractParameters: Parameters = { resourceType: "Parameters", parameter: [] };
+  public extractServiceUrl: string = "https://fhir.forms-lab.com/QuestionnaireResponse/$extract";
+  public extractingInProgress: boolean = false;
 
   // Internal variables
   expressionEditor?: ace.Ace.Editor | undefined = undefined;
   public internalValue: string | undefined = undefined;
 
   public setValue(value: string | undefined): void {
+    console.log("setValue called");
     this.internalValue = value;
     this.ensureEditorIsCreated();
-    this.expressionEditor?.setValue(value ?? "");
+    // this.expressionEditor?.setValue(value ?? "");
     // this.expressionEditor?.clearSelection();
   }
 
-  public getValue(): string | undefined {
-    // if (!this.expressionEditor)
-      return this.internalValue;
-    
-    // return this.expressionEditor?.getValue();
+  async performExtractOperation(): Promise<void> {
+    console.log("Extracting the questionnaire");
+
+    // update the parameters object for the extract
+    let typeBasedExtract = this.extractServiceUrl.endsWith("/QuestionnaireResponse/$extract");
+    this.extractParameters.parameter = [
+      {
+        name: "questionnaire-response",
+        resource: this.questionnaireResponse
+      }
+    ];
+    if (typeBasedExtract) {
+      this.extractParameters.parameter.push({
+        name: "type",
+        valueString: "Questionnaire"
+      });
+    }
+    this.expressionEditor?.setValue('');
+    this.expressionEditor?.clearSelection();
+
+    // Now pass this to the extract service
+    this.extractingInProgress = true;
+
+    // Send the extract request
+    try
+    {
+      const response = await fetch(this.extractServiceUrl, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+          "Accept": "application/json"
+        },
+        body: JSON.stringify(this.extractParameters)
+      });
+      if (response.ok)
+      {
+        const extractResponse = await response.json();
+        this.expressionEditor?.setValue(JSON.stringify(extractResponse, null, 2));
+        this.expressionEditor?.clearSelection();
+        console.log("Extracted response", extractResponse);
+      }
+      else
+      {
+        console.error("Failed to extract the questionnaire", response);
+      }
+    }
+    catch (error)
+    {
+      console.error("Error extracting the questionnaire", error);
+    }
+    this.extractingInProgress = false;
+  }
+
+  @Watch('questionnaireResponse', { immediate: false, deep: true })
+  async onQuestionnaireResponseChanged() {
+    console.log("QR changed");
+  }
+
+  @Watch('questionnaire', { immediate: false, deep: true })
+  async onQuestionnaireChanged() {
+    console.log("Q changed");
   }
 
   ensureEditorIsCreated(): void {
     try {
       var editorDiv: any = this.$refs.aceEditorExpression as HTMLDivElement;
       if (editorDiv && this.expressionEditor === undefined) {
-        console.log("need to create the editor");
+        console.log("creating the extract ace editor");
         this.expressionEditor = ace.edit(editorDiv, {
           wrap: "free",
           minLines: 3,
-          maxLines: Infinity,
+          maxLines: 35,
           highlightActiveLine: false,
-          showGutter: false,
-          fontSize: 16,
+          showGutter: true,
+          fontSize: 14,
           cursorStyle: "slim",
           showPrintMargin: false,
           theme: "ace/theme/chrome",
-          mode: "ace/mode/text",
+          mode: "ace/mode/json",
           wrapBehavioursEnabled: true
         });
-         console.log("extract initialized");
+        console.log("extract ace editor  initialized");
 
-        if (this.expressionEditor){
+        if (this.expressionEditor) {
           setCustomHighlightRules(this.expressionEditor, FhirPathHightlighter_Rules);
-          this.expressionEditor.setValue(this.internalValue ?? "");
-          this.expressionEditor.clearSelection();
+          // this.expressionEditor.setValue(this.internalValue ?? "");
+          // this.expressionEditor.clearSelection();
           // this.expressionEditor.on("change", this.fhirpathExpressionChangedEvent)
           console.log("editor configured");
         }
-        else{
+        else {
           console.error("Failed to create the editor");
         }
       }
@@ -106,10 +177,9 @@ export default class QuestionnaireExtractTest extends Vue {
 
   updated(): void {
     this.ensureEditorIsCreated();
-    this.internalValue = JSON.stringify(this.questionnaire, null, 2);
     console.log('updated');
     this.$nextTick(() => {
-      if (this.expressionEditor){
+      if (this.expressionEditor) {
         this.expressionEditor.renderer.updateFull(true)
         this.expressionEditor.resize();
       }